Turn lizp.h into an almost complete single-file header library
